+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Change Log
 
-### 1.2.x
+### 1.2.2
 
 * Fixed:    Fixed missing transforms when using the MysqlCursor!
 * Fixed:    Collection::calculateAggregates() returns nulls if called on an empty collection